% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_filltemplates_helpers_02.R
\name{armamod_whf}
\alias{armamod_whf}
\title{Generate ARMA Model as (a(z), b(z))}
\usage{
armamod_whf(theta, tmpl)
}
\arguments{
\item{theta}{vector of doubles; deep parameters}

\item{tmpl}{template as obtained by \link{tmpl_whf_rev}, i.e. a list containing slots \code{polm_ar}, \code{polm_ma_bwd}, \code{polm_ma_fwd}, \code{B}.}
}
\value{
List with slots \code{polm_ar}, \code{polm_ma}, (both \link{polm} objects), and \code{B} (a matrix).
}
\description{
Called when generating IRFs.
Calls \link{fill_tmpl_whf_rev}.
\cr
Takes the deep parameters \code{theta} as input and returns an ARMA model (not in normalised canonical WHF form) 
by first filling the ARMA-WHF template \code{tmpl} and subsequent transformation.
}
