% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_ratmat_tools.R
\name{array2data.frame}
\alias{array2data.frame}
\title{Coerce arrays to data frames}
\usage{
array2data.frame(x, rows = NULL, cols = NULL)
}
\arguments{
\item{x}{Matrix or array.}

\item{rows, cols}{integer vectors. These two vectors define a partition of the "dimensions" (1,...,n),
where n is the number of dimensions of x (i.e. length(dim(x))).
If either of the two is missing, then the complement is used.
At least one of the arguments "rows" and "cols" has to be given.}
}
\value{
data.frame
}
\description{
This function was originally part of the R-package \strong{rationalmatrices}.
\cr
This helper function creates a \code{data.frame} from a given array.
}
\examples{
# test array
x = test_array(dim = c(2,3,2), dimnames = TRUE)
array2data.frame(x, cols = c(1,3,2))
array2data.frame(x, rows = 1)
array2data.frame(x, rows = 2:1)
array2data.frame(x, rows = c(2,1,3))

# consider a pseudo socio economic data set
x = test_array(dim = c(2,4,5), random = TRUE)
dimnames(x) = list(sex=c('female','male'),
                   education = c('none','primary','high','university'),
                   age = c('<20','30-40','40-50','50-60','>60'))
array2data.frame(x, cols = 1)
array2data.frame(x, cols = c(1,2))

}
\seealso{
The helper function \code{array2data.frame} is used
         internally for the \code{\link{plot methods}}. 
         The function \code{\link{bmatrix}} (which coerces arrays to matrices) is a simplified
         version of \code{array2data.frame}.
}
