% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_rldm_print_methods.R
\name{as_tex_matrix}
\alias{as_tex_matrix}
\title{TeX Matrix}
\usage{
as_tex_matrix(x)
}
\arguments{
\item{x}{matrix, where \code{paste(x[i,j])} returns a valid "TeX" string.}
}
\value{
character string.
}
\description{
This function was originally part of the R-package \strong{RLDM}.
}
\examples{
as_tex_matrix(diag(1:2, nrow = 2, ncol = 3))
}
