% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_rldm_print_methods.R
\name{as_tex_matrixpoly}
\alias{as_tex_matrixpoly}
\title{TeX Matrix Polynomials}
\usage{
as_tex_matrixpoly(coefs, x = "z", as_matrix_of_polynomials = TRUE)
}
\arguments{
\item{coefs}{3-dimensional array with the coefficients of the matrix polynomial.}

\item{x}{(character string) polynomial variable or process variable.}

\item{as_matrix_of_polynomials}{boolean.}
}
\value{
character string.
}
\description{
This function was originally part of the R-package \strong{RLDM}.
}
\examples{
coefs = array(round(rnorm(2*3*1), 1), dim = c(2,3,2))

as_tex_matrixpoly(coefs)
as_tex_matrixpoly(coefs, x = 'x', as_matrix_of_polynomials = FALSE)
}
