% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_rldm_autocov_methods.R
\name{autocov}
\alias{autocov}
\alias{autocov.default}
\title{Autocovariance, Autocorelation and Partial Autocorrelation Function}
\usage{
autocov(obj, type, ...)

\method{autocov}{default}(
  obj,
  type = c("covariance", "correlation", "partial"),
  lag.max = NULL,
  na.action = stats::na.fail,
  demean = TRUE,
  ...
)
}
\arguments{
\item{obj}{either a \code{\link{armamod}}, \code{\link{stspmod}}, \code{\link{autocov}} object
or a "data" object.}

\item{type}{character string giving the type of acf to be computed. Allowed values are
"covariance" (the default), "correlation", or "partial". Will be partially matched.
Note that the default value here is "covariance" whereas \code{\link[stats]{acf}}
uses "correlation" as default.}

\item{...}{not used.}

\item{lag.max}{(integer) maximum lag.}

\item{na.action}{function to be called to handle missing values. \code{\link[stats]{na.pass}} can be used.}

\item{demean}{logical. Should the covariances be about the sample means?}
}
\value{
\code{autocov} object, i.e. a list with slots
\item{acf}{\code{\link[rationalmatrices]{pseries}} object, which stores the covariances (correlations).}
\item{type}{character string which indicates the type of the ACF.}
\item{gamma}{(m,m,lag.max+1) dimensional array which stores the autocovariance function.}
\item{names}{(m)-dimensional character vector or NULL. This optional slot stores the names
             for the components of the time series/process.}
\item{label}{character string or NULL.}
\item{n.obs}{integer or NULL. This slot stores the sample size.}
}
\description{
This function was originally part of R package \strong{RLDM}.
\cr
Compute respectively estimate the autocovariance, autocorrelation or partial autocorrelation function of a stationary process.
}
\details{
The class of the input parameter "\code{obj}" determines the S3 method called and hence what is actually computed.

\bold{Population ACF:}

If "\code{obj}" is an \code{\link{armamod}} or \code{\link{stspmod}} object then \code{autocov(obj, ...)} computes the ACF of the corresponding stationary process.

Note however, that the function returns nonsense, if the model does not satisfy the \emph{stability} condition.

\bold{Change the type of an ACF:}

Calling \code{autocov(obj, type)}, where "\code{obj}" is an \code{autocov} object returns an ACF of the desired type.
E.g. if "\code{obj}" holds a partial autocorrelation function then \code{autocov(obj, type = 'covariance')} may be used to retrieve the corresponding autocovariance function.

This is possible since the \code{autocov} object stores the "original" autocovariances in a slot named \code{gamma}.

\bold{Sample ACF:}

The default S3 method estimates the ACF from given data.
It assumes that "\code{obj}" is a univariate or multivariate numeric time series object,
a (numeric) data frame or a (numeric) vector, respectively matrix
and then simply calls the function \code{\link[stats]{acf}} in the \pkg{stats} package to compute the sample autocovariance function.
If needed, then the corresponding sample autocorrelation, respectively sample partial autocorrelation function is computed (and returned).

The syntax is quite analogous to \code{\link[stats]{acf}}, so please consider the documentation of \code{\link[stats]{acf}} for more details.

Note that \pkg{stats} stores autocovariance/autocorrelation functions as \code{(lag.max+1,m,m)} dimensional arrays, whereas \pkg{RLDM} uses \code{(m,m,lag.max+1)} dimensional arrays.

The definition of partial autocorrelations used by \code{\link[stats]{acf}} differs from the definition used here.
Furthermore \code{\link[stats]{acf}} skips the lag zero partial autocorrelation coefficient
and thus the pacf computed by  \code{\link[stats]{acf}} is (lag.max,n,n) dimensional.

The default choice for the number of lags is \eqn{10*log10(N/m)} where \eqn{N}
is the number of observations and \eqn{m} the number of series.
This number will be automatically limited to one less than the number
of observations in the series.
}
\seealso{
The autocovariance function of (V)ARMA processes may also be
computed by \code{\link[stats]{ARMAacf}} in the scalar case and by
\code{\link[MTS]{VARMAcov}} in the multivariate case (m > 1).

As noted above the sample ACF is computed via the \code{\link[stats]{acf}} routine in the \pkg{stats} package.
}
