% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_ratmat_tools.R
\name{bdiag}
\alias{bdiag}
\title{Block Diagonal Matrix}
\usage{
bdiag(...)
}
\arguments{
\item{...}{matrices or vectors. Vectors are treated as diagonal matrices.
If no input arguments are provided then \code{bdiag} returns \code{NULL}.}
}
\value{
block diagonal matrix (or \code{NULL} if \code{bdiag} has been called without inputs).
}
\description{
This function was originally part of the R-package \strong{rationalmatrices}.
\cr
Combine two or more matrices to a block-diagonal matrix. The functions supports boolean, integer,
numeric and complex matrices (and vectors). The procedure makes some effort to retain the
(col- / row-) names of the inputs.
}
\examples{
A = matrix(TRUE, nrow = 2, ncol = 3)
colnames(A) = paste('A',1:3,sep ='.')
B = rep(2L,3)
names(B) = paste('B',1:3,sep = '.')
C = matrix(0, nrow = 3, ncol = 0)
rownames(C) = paste('C',1:3)
D = matrix(4, nrow = 2, ncol = 3)
E = matrix(complex(real=5), nrow = 2, ncol = 2)
rownames(E) = paste('E',1:2,sep='.')
colnames(E) = paste('E',1:2,sep='.')

bdiag()
bdiag(NULL,NULL)

X = bdiag(A,NULL) # NULL arguments are skipped
X
str(X)            # output is of type 'logi'

X = bdiag(A,NULL,B)
X
str(X)            # output is of type 'int'

# note the action of the "empty" (3 times 0) matrix C
X = bdiag(A,C,B)
X
str(X)            # output is of type 'num'

X = bdiag(A,B,C,D,E)
X
str(X)            # output is of type 'cplx'

\dontrun{
# the inputs must be vectors or matrices. arrays are not supported.
bdiag(A, array(1, dim = c(2,3,1)))

# character matrices are problematic, since it is not clear how to set the
# non diagonal elements. Therefore, the following statement throws an error.
bdiag(A, matrix('B', nrow = 2, ncol = 1), 3:4)
}
}
