% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_ratmat_tools.R
\name{bhankel}
\alias{bhankel}
\title{Block Hankel matrix}
\usage{
bhankel(R, d = NULL)
}
\arguments{
\item{R}{3-dimensional array, matrix or vector. A vector of length \eqn{k} is coerced to a
\eqn{(1,1,k)}-dimensional array and a \eqn{(p,q)} matrix is treated as an array of
dimension \eqn{(p,q,1)}.}

\item{d}{determines the number of block rows and columns.
Suppose that \code{R} is an array of size \eqn{(p,q,k)}.
If \eqn{d=(m,n)} then \code{bhankel} returns a block Hankel matrix with \eqn{m} block rows and \eqn{n} block columns.
If \eqn{d=m} then a Hankel matrix with \eqn{m} block rows and
\eqn{n=\max(k+1-m,1)}{n=max(k+1-m,1)} block columns is returned.
In the default case \code{d = NULL} the number of block rows is \eqn{m=(k+1)/2} for odd \eqn{k}
and \eqn{m=(k/2+1)} else and the number of block columns is set to \eqn{n=\max(k+1-m,1)}{n=max(k+1-m,1)}.
In the case \eqn{(m+n-1)>k} the array \code{R} is padded with zeros.}
}
\value{
Matrix of size \eqn{(pm\times qn)}{(pm x qn)}.
}
\description{
This function was originally part of the R-package \strong{rationalmatrices}.
\cr
Construct a block Hankel matrix with (block) entries from a 3-dimensional array R.
The \eqn{(i,j)}-th block of the Hankel matrix is \code{R[,,i+j-1]}.
}
\examples{
bhankel(1:6)
bhankel(1:6, d = 3)
bhankel(letters[1:6], d = c(3,7))  # note the "zero padding"
bhankel(test_array(dim = c(2,2,6)))
bhankel(test_array(dim = c(3,2,6)), d = 3)
bhankel(test_array(dim = c(2,2,6)), d = c(3,7))
bhankel(test_array(dim = c(1,2,6)), d = c(1,2))
bhankel(test_array(dim = c(2,3)), d = c(2,2))
bhankel(test_array(dim = c(2,2,6)), d = c(3,0))
bhankel(test_array(dim = c(2,2,0)), d = c(2,2))

\dontrun{
# the following examples throw an error
bhankel(1:5, d = c(-1,1))
bhankel(test_array(dim = c(2,3,2,1)))
}
}
