% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_ratmat_tools.R
\name{bmatrix}
\alias{bmatrix}
\title{Block matrices}
\usage{
bmatrix(x, rows = NULL, cols = NULL)
}
\arguments{
\item{x}{Vector, matrix or array. Vectors are coerced to one dimensional arrays and matrices
are of course treated as 2-dimensional arrays.}

\item{rows, cols}{integer vectors. These two vectors define a partition of the "dimensions" \code{(1,...,n)}, where
\code{n} is the number of dimensions of \code{x} (i.e. \code{length(dim(x))}).
If either of the two is missing, then the complement is used. At least one of the arguments "rows" and
"cols" has to be given.}
}
\value{
matrix
}
\description{
This function was originally part of the R-package \strong{rationalmatrices}.
\cr
This helper function coerces an array to a (block) matrix.
}
\examples{
x = 1:4
bmatrix(x, rows = 1, cols = integer(0)) # returns an (4,1) matrix
bmatrix(x, cols = 1, rows = NULL)       # returns an (1,4) matrix

x = test_array(dim = c(2,3))
bmatrix(x, cols = 2)    # returns x    (is equivalent to bmatrix(x, rows = 1, cols = 2))
bmatrix(x, rows = 2)    # returns t(x) (is equivalent to bmatrix(x, rows = 2, cols = 1))
bmatrix(x, rows = integer(0))   # returns an (1,6) matrix

x = test_array(dim = c(2,3,4))
bmatrix(x, rows = 1)   # is equivalent to: bmatrix(x, rows = 1, cols = c(2,3))
bmatrix(x, rows = 1, cols = c(3,2))
bmatrix(x, cols = 2)   # is equivalent to: bmatrix(x, rows = c(1,3), cols = 2)
bmatrix(x, rows = 1:3) # is equivalent to: bmatrix(x, cols = integer(0))
bmatrix(x, rows = c(3,1,2))

\dontrun{
# the examples below throw an error
bmatrix(x, rows = 1, cols = 2)
bmatrix(x, rows = c(1,2), cols = c(2,3))
bmatrix(x, rows = c(1,2,1), cols = 3)
}
}
