% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_ratmat_plot_tools.R
\name{default_colmap}
\alias{default_colmap}
\title{Default Color Palette}
\usage{
default_colmap(n)
}
\arguments{
\item{n}{integer}
}
\value{
(character) vector of length \code{n} with color codes (\code{"#RRGGBB"}).
}
\description{
Default color palette.
}
\examples{
default_colmap(3)
}
\keyword{internal}
