% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_rldm_freqresp_methods.R
\name{dft_3D}
\alias{dft_3D}
\title{Discrete Time Fourier Transform}
\usage{
dft_3D(a, n.f = dim(a)[3])
}
\arguments{
\item{a}{\eqn{(m,n,k)} dimensional (numeric) array}

\item{n.f}{(integer) number of frequencies}
}
\value{
A \code{\link[rationalmatrices]{zvalues}} object.
}
\description{
Compute the Discrete Time Fourier Transform for data
stored in a 3-dimensional array.
}
\keyword{internal}
