% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_ratmat_extract_methods.R
\name{extract}
\alias{extract}
\alias{[.polm}
\alias{[.lmfd}
\alias{$.lmfd}
\title{Extract Parts of a Rational Matrix}
\usage{
\method{[}{polm}(x, i, j)

\method{[}{lmfd}(x, i, j)

\method{$}{lmfd}(x, name)
}
\arguments{
\item{x}{a rational matrix, i.e. a \code{\link{polm}}, \code{\link{lmfd}}, \code{\link{stsp}}, 
\code{\link{pseries}} or \code{\link{zvalues}} object.}

\item{i, j}{indices (integer or boolean vector)}

\item{name}{character (A,B,C,D for statespace objects and a,b for LMFD objects)}
}
\value{
rational matrix
}
\description{
This function was originally part of the R-package \strong{rationalmatrices}.
\cr
The subsetting operation \code{x[,]} for rational matrices works analogously to the subsetting of 
ordinary matrices. However, this operator is not implemented for \code{\link{lmfd}} objects. 
(x[i,j] throws an error if x is an \code{lmfd} object!) The result is an object of the same class. 
\cr
The \code{$} operator may be used to extract the \eqn{a}, \eqn{b} polynomial of a left matrix 
fraction description (\code{\link{lmfd}} object) and the matrices 
\eqn{A,B,C,D} of a state space representation (\code{\link{stsp}} object).
}
\details{
\itemize{
\item \code{x[]} or \code{x[,]} simply return the original object.
\item \code{x[i]} returns a "vector", i.e. an \eqn{(s,1)} dimensional matrix. 
\item \code{x[i,]},\code{x[,j]} or \code{x[i,j]} return a rational matrix with rows selected by \code{i} 
      and columns selected by \code{j}. 
}
Note that "named" arguments are not supported (in order to simplify the coding). 

In order to have a finer control, one may e.g. use \code{unclass(x)[,,]}.
}
\examples{
# polynomial matrices 
a = test_polm(dim = c(3,2), degree = 1)
a[]          # returns a
a[,]         # returns a
a[c(1,3,6)]  # returns a "vector" with the (1,1), (3,1) and (3,2) element of a
a[1,]        # returns the first row of a 
a[,2]        # returns the second column of a 
a[c(TRUE,FALSE,TRUE),c(FALSE, TRUE)] # returns a 2 by 1 matrix 
a[c(1,1),c(2,1)] # returns a 2 by 2 matrix 

\dontrun{
# the subsetting operator [,] is not implemented for "lmfd" objects
test_lmfd(dim = c(2,2), degrees = c(1,1))[1,1]

a[i=1, j=2] # throws an error, since "named" arguments are not allowed.
}
}
