% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_llfactory.R
\name{extract_deep_rev}
\alias{extract_deep_rev}
\title{Extract deep parameters from SARMA-WHF model given a template}
\usage{
extract_deep_rev(armamod_whf_obj, tmpl, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{armamod_whf_obj}{List of 3D arrays \code{polm_ar}, \code{polm_ma_bwd}, \code{polm_ma_fwd}, and the static shock transmission matrix \code{B}, 
as well as info regarding the distribution in slot \code{distr} Obtained from \link{fill_tmpl_whf_rev}.}

\item{tmpl}{List. Affine mappings etc specifiying a SVARMA-WHF template. See return value of \link{tmpl_whf_rev}.}

\item{tol}{Tolerance level. "Residuals" when extracting deep parameters must be smaller than this.
Otherwise an error is thrown.
Default equal to \code{sqrt(.Machine$double.eps)}}
}
\value{
Vector of doubles containing deep parameters.
}
\description{
Called by \link{get_init_armamod_whf_random}.
\cr
Uses the affine mappings in \code{tmpl} to reconstruct the deep parameters from the linear parameters in \code{armamod_whf_obj} with \link[stats]{lsfit}.
}
