% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_filltemplates_helpers_02.R
\name{f2g}
\alias{f2g}
\alias{g2f}
\title{Transform forward polynomial of WHF to polynomial in z}
\usage{
f2g(f, kappa_k)

g2f(g, kappa_k)
}
\arguments{
\item{f}{Array object which contains f(z).
There is no class for polynomials in 1/z so far.}

\item{kappa_k}{Two dimensional vector of integers parametrising the partial indices of the WHF}

\item{g}{\link{polm} object which contains \eqn{g(z) = s(z)f(z)}.}
}
\value{
\link{polm} object \eqn{g(z)} or array object for \eqn{f(z)}.
}
\description{
Called in \link{armamod_whf}.
\cr
Given the partial indices, parametrised by \eqn{\kappa} and \code{k},
we transform \eqn{f(z)} (a polynomial in \eqn{\frac{1}{z}}) to \eqn{g(z) = s(z)f(z)} and vice versa,
see the paper for notation.
}
