% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_filltemplates_helpers_02.R
\name{fill_tmpl_whf_rev}
\alias{fill_tmpl_whf_rev}
\title{Fill WHF template with deep parameters}
\usage{
fill_tmpl_whf_rev(theta, tmpl)
}
\arguments{
\item{theta}{Vector of doubles. Deep parameters to be filled in the SVARMA-WHF template \code{tmpl}.}

\item{tmpl}{SVARMA-WHF template, i.e. list with slots containing affine mappings between deep and linear parameters, see return value of \link{tmpl_whf_rev}.}
}
\value{
List of 3D arrays \code{polm_ar}, \code{polm_ma_bwd}, \code{polm_ma_fwd}, 
  and the static shock transmission matrix \code{B}, 
  as well as info regarding the distribution in slot \code{distr}
}
\description{
Called in \link{ll_whf_factory}, \link{armamod_whf}.
}
\details{
Use affine mappings in SVARMA-WHF template to fill free parameters in AR, MA bwd, etc objects (3D arrays) with deep parameters.
}
