% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_llfactory.R
\name{get_ic}
\alias{get_ic}
\title{Rotate inputs to Independence}
\usage{
get_ic(res)
}
\arguments{
\item{res}{Residuals in long form}
}
\value{
Static shock transmission matrix \eqn{B = C M} where \eqn{C} is obtained from 
  Cholesky decomposition of \code{1/n_obs t(res) res = C t(C)} and 
  \code{M} is the mixing matrix obtained from the output of \link{steadyICA} applied to \eqn{res solve(t(C))}.
}
\description{
Called in \link{create_results_list}.
\cr
Uses \link[steadyICA]{steadyICA} in order to obtain independent components.
This function is called after the initial Gaussian estimation and should serve as initial estimate for the noise parameters.
Afterwards, full ML with Laplace density and SGT densities is performed.
}
