% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_template_initvals.R
\name{get_init_armamod_whf_random}
\alias{get_init_armamod_whf_random}
\title{Obtain Initial deep parameters}
\usage{
get_init_armamod_whf_random(data_long, tmpl)
}
\arguments{
\item{data_long}{Matrix. Contains observation in each row, number of variables corresponds to number of rows. 
Note that the data need to be \strong{demeaned}. This is not checked!}

\item{tmpl}{SVARMAWHF template. As obtained from \link{tmpl_whf_rev}.}
}
\value{
Vector of deep system and noise parameters. 
  They are used to fill a given template.
}
\description{
Called by main script \code{scriptsP_whf_revision_sgt_bq/aa_Rscript4slurm_bq.R}.
Calls \link{est_ar}, \link{extract_deep_rev}, \link{get_residuals}.
\cr
Obtains initial deep parameters for given data \code{data_long} and a given SVARMAWHF template \code{tmpl} obtained from \link{tmpl_whf_rev}.
The AR parameters are obtained from the Yule-Walker equations, see \link{est_ar}. 
In order to ensure that the initial value for the backward and forward MA polynomial matrix be stable (have its determinantal roots on the correct side of the unit circle),
we initial them with the (possibly randomized) Minnesota prior \link[LaplacesDemon]{MinnesotaPrior}.
\cr
Subsequently, we obtain the residuals for the template filled with the initial deep parameter values via \link{get_residuals}.
}
\section{Note}{

Case \code{q = p} throws an error because otherwise problems with \link[stats]{lsfit} in \link{extract_deep_rev} would occur. 
Handling in the code would be nicer than in the Rmarkdown file...
}

