% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_residuals}
\alias{get_residuals}
\title{Obtain residuals with Rcpp}
\usage{
get_residuals(data_in, polm_ar, polm_ma_bwd, polm_ma_fwd, kappa, k)
}
\arguments{
\item{data_in}{const arma::mat reference. Observations correspond to columns.}

\item{polm_ar, polm_ma_fwd}{const arma::mat references. Wide matrices.
\eqn{I_n, a_1, \ldots, a_p} and \eqn{f_1, \ldots, f_K}}

\item{polm_ma_bwd}{arma::mat reference. 
\strong{NOT CONST} because \eqn{p_0} needs to be inverted, 
applied to \eqn{p_q, \ldots, p_1, p_0}, and then the new \eqn{p_0 = I_n} needs to be discarded}

\item{kappa, k}{const arma::uword references. Partial indices}
}
\value{
Object of type arma::mat. Wide data matrix (rows correspond to variables, observations to columns) of residuals.
}
\description{
For given AR, bwd MA, fwd MA parameters, calculate the residuals (B and Sigma are taken into account later).
The input arguments \code{polm_ar}, \code{polm_ma_fwd}, \code{polm_ma_bwd} are in a different format than usual for this Rcpp function, 
i.e. they are in wide matrices.
Also note that \eqn{p_0} is not necessarily the identity matrix and needs to be inverted in the code below.
\cr
What is happening in this function is better documented in its R version: \link{get_residuals_R}
}
