% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_ratmat_plot_tools.R
\name{get_subfigure_layout}
\alias{get_subfigure_layout}
\title{Subfigure Layout}
\usage{
get_subfigure_layout(margins)
}
\arguments{
\item{margins}{A named list with slots \code{bottom}, \code{left}, \code{top}
and \code{right} which determine the margins of the sub-figures.
E.g. \code{bottom} is an \code{m}-dimensional vector where
\code{bottom[i]} gives the bottom margin (in "line units")
of the sub-figures in the \code{i}-th row.}
}
\value{
A list with slots
\item{omi}{4-dimensional vector with the "outer margin" in inches.}
\item{mar}{(m,n,4)-dimensional array, where \code{mar[i,j,]} contains the
           margins (in "line" units) of the (i,j)-th sub-figure.}
\item{fig}{(m,n,4)-dimensional array, where \code{fig[i,j,]} contains the
           coordinates (in "NDC" units) of the figure region of the
           (i,j)-th sub-figure.}
}
\description{
The device is split up into an (m-by-n) array of (sub-) figures. This tool
computes the corresponding figure region and margins for each of the
(sub-) figures.
}
\seealso{
\code{\link[graphics]{par}} for more details.
}
\keyword{internal}
