% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_ratmat_arithmetic_methods.R
\name{\%r\%}
\alias{\%r\%}
\title{Matrix Multiplication of Rational Matrices}
\usage{
e1 \%r\% e2
}
\arguments{
\item{e1, e2}{two rational matrix objects (i.e. \code{\link{polm}}, \code{\link{lmfd}}, \code{\link{rmfd}},
\code{\link{stsp}}, \code{\link{pseries}} or \code{\link{zvalues}} objects), 
or objects which may be coerced to rational matrix objects.}
}
\value{
rational matrix object. The class depends on the classes of the arguments \code{e1,e1}.
}
\description{
This function was originally part of the R-package \strong{rationalmatrices}.
}
