% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_ratmat_tools.R
\name{idx_trafo}
\alias{idx_trafo}
\alias{ind2sub}
\alias{sub2ind}
\title{Transform between Linear Index and Matrix Indices}
\usage{
ind2sub(dim, ind)

sub2ind(dim, row, col)
}
\arguments{
\item{dim}{Integer vector of size 2. Matrix dimensions.}

\item{ind}{Integer. Linear index.}

\item{row}{Integer. Row index.}

\item{col}{Integer. Column index.}
}
\value{
\code{ind2sub()} returns for given linear index, a matrix index (row, col).
  \code{sub2ind()} returns for a given matrix index (row, col), a linear index (column-major).
}
\description{
This function was originally part of the R-package \strong{rationalmatrices}.
\cr
These functions are as in MATLAB.
\code{ind2sub()} transforms a linear index to a row and column index for a matrix of given size.
\code{sub2ind()} transforms a matrix index, (row, col) to a a linear index (in terms of columns).
}
\examples{
A = matrix(1:(3*4), 3, 4)
A

ind2sub(c(3,4), 7)
sub2ind(c(3,4), 2, 3)
}
