% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_filltemplates_helpers_02.R
\name{irf_change}
\alias{irf_change}
\title{Permute or change sign of shocks to IRF}
\usage{
irf_change(pseries, perm = NULL, sign = NULL)
}
\arguments{
\item{pseries}{pseries or array object}

\item{perm}{vector of dimension \code{dim_in}, describing the permutation to be postmultiplied on the transfer function}

\item{sign}{vector of dimension \code{dim_in} containing -1 and 1 as elements}
}
\value{
pseries object
}
\description{
First the permutation is right-multiplied on the IRF, then the sign matrix.
}
\examples{
ps = test_polm(dim = c(2,2), degree = 10, random = TRUE) \%>\% pseries()
irf_change(ps, c(2,1), c(1,-1))
}
