% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_filltemplates_helpers_02.R
\name{irf_unique}
\alias{irf_unique}
\title{Unique (rotated) IRF}
\usage{
irf_unique(theta, tmpl, lag.max = 30)
}
\arguments{
\item{theta}{doubles; deep parameters}

\item{tmpl}{SVARMA-WHF template, see \link{tmpl_whf_rev}}

\item{lag.max}{integer; maximal number of lags to be contained in the output
Default set to 30.}
}
\value{
\link{pseries} object. IRF with 30 lags.
}
\description{
Fixes permutations and signs of IRF as described in the article.
}
