% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_filltemplates_helpers_02.R
\name{irf_whf}
\alias{irf_whf}
\title{Generate IRF}
\usage{
irf_whf(theta, tmpl, n_lags)
}
\arguments{
\item{theta, tmpl}{Used in call to \link{fill_tmpl_whf_rev}.}

\item{n_lags}{Number of lags for IRF}
}
\value{
\code{pseries} object
}
\description{
Called in analysis when comparing IRFs.
Calls \link{armamod_whf}.
}
