% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_ratmat_is_methods.R
\name{is.class methods}
\alias{is.class methods}
\alias{is.polm}
\alias{is.lmfd}
\title{Check Objects}
\usage{
is.polm(x)

is.lmfd(x)
}
\arguments{
\item{x}{object}
}
\value{
\code{TRUE} if the argument is a valid object.
}
\description{
This function was originally part of the R-package \strong{rationalmatrices}.
\cr
Check if the argument is a valid object.
}
