% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_ratmat_tools.R
\name{iseq}
\alias{iseq}
\title{Sequence generation}
\usage{
iseq(from = 1, to = 1)
}
\arguments{
\item{from, to}{the starting and end values of the sequence.}
}
\value{
\code{iseq} returns an empty integer vector if \code{to} is less than \code{from} and
   \code{seq(from,to)} else.
}
\description{
This function was originally part of the R-package \strong{rationalmatrices}.
\cr
}
\details{
As opposed to the standard \code{\link{seq}} command the function \code{iseq} returns an empty
vector if the starting value is larger than the end value.

More general than \link{seq_len} because the sequence does not need to start from 1.
}
\examples{
iseq(0,1) # => c(0,1)
iseq(1,0) # => integer(0)
seq(1,0)  # => c(1,0)
}
