% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_ratmat_plot_tools.R
\name{line2inch}
\alias{line2inch}
\title{Convert 'line units' to inches}
\usage{
line2inch()
}
\value{
scalar
}
\description{
Convert 'line units' to inches
}
\examples{
line2inch()
par(cex = 2, mex = 2)
line2inch()
}
\keyword{internal}
