% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_llfactory.R
\name{ll_whf_factory}
\alias{ll_whf_factory}
\title{Function factory for negative log-likelihood minimization}
\usage{
ll_whf_factory(
  data_wide,
  tmpl,
  shock_distr = c("laplace", "mixture", "gaussian", "sgt", "skewed_t", "skewed_ged"),
  noise_only = FALSE,
  use_cpp = TRUE,
  info = FALSE
)
}
\arguments{
\item{data_wide}{Matrix. Data in wide form, i.e. every column contains an observation.}

\item{tmpl}{List. Containing the mappings from deep to linear parameters, see return value of \link{tmpl_whf_rev}.}

\item{shock_distr}{Character. One of \code{laplace}, \code{mixture}, \code{gaussian}, \code{sgt}, \code{skewed_t}, \code{skewed_ged}}

\item{noise_only}{boolean. Default set to \code{FALSE}. If \code{TRUE}, then in the likelihood function to be created only the noise parameters are optimized (it depends only on the noise parameters in this case).}

\item{use_cpp}{Boolean. Default TRUE. If FALSE, then the R version is used instead of the faster Rcpp version.
Used for debugging and checking.}

\item{info}{Boolean. Default set to \code{FALSE}. 
Only used for debugging in which case tibble_info and call_counter must be in the global environment.}
}
\value{
Closure for particular SVARMA-WHF template \code{tmpl} and shock distribution \code{shock_distr} which depends only on deep parameters.
}
\description{
Called by \link{create_results_list}.
Calls \link{fill_tmpl_whf_rev}, \link{get_residuals}.
\cr
Creates a closure (function plus enclosing environment) which depends only on the deep parameters.
In this way, for each template, i.e. integer-valued parameters, and each shock distribution, a log-likelihood function can be created which also contains the data set.
This is convenient for optimisation via \link[stats]{optim}.
\cr
In the main script \code{scriptsP_whf_revision_sgt_bq/aa_Rscript4slurm_bq.R}, for each tuple of integer-valued parameters \eqn{(p,q,\kappa,k)} a log-likelihood function for each shock distribution \code{shock_distr} is created.
}
