% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_ll_gradient.R
\name{ll_whf_factory_grad_helper}
\alias{ll_whf_factory_grad_helper}
\title{Function factory for gradient of log-likelihood function}
\usage{
ll_whf_factory_grad_helper(
  data_wide,
  tmpl,
  use_cpp = TRUE,
  shock_distr = "sgt"
)
}
\arguments{
\item{data_wide}{Matrix. Data in wide form, i.e. every column contains an observation.}

\item{tmpl}{List. Containing the mappings from deep to linear parameters, see return value of \link{tmpl_whf_rev}.}

\item{use_cpp}{Boolean. Default TRUE. If FALSE, then the R version is used instead of the faster Rcpp version.
Used for debugging and checking.}

\item{shock_distr}{Default \code{sgt}. Other option is \code{skewed_t}}
}
\value{
Closure for particular SVARMA-WHF template \code{tmpl} using the SGT distribution.
  Depends only on deep parameters, the dataset is contained in the enclosing environment.
}
\description{
Not used for optimisation but rather obtaining the OPG version of the information matrix.
\cr
Creates a closure (function plus enclosing environment) which depends only on the deep parameters.
In this way, for each template, i.e. integer-valued parameters, and each shock distribution, a log-likelihood function can be created which also contains the data set.
It uses similar code as \link{ll_whf_factory} but only exists for the SGT and the skewed t distribution.
}
