% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_ratmat_plot_methods.R
\name{plot methods}
\alias{plot methods}
\alias{plot.pseries}
\alias{plot.zvalues}
\title{Plot Methods}
\usage{
\method{plot}{pseries}(
  x,
  x_list = NULL,
  xlim = c("global", "column", "subfig"),
  ylim = c("row", "subfig", "global"),
  main = "impulse response",
  xlab = "lag (k)",
  ylab = NULL,
  subfigure_main = NA,
  parse_subfigure_main = FALSE,
  style = c("gray", "bw", "bw2", "colored"),
  col = NA,
  type = "l",
  lty = "solid",
  lwd = 1,
  pch = 16,
  cex.points = 1,
  bg.points = "black",
  legend = NULL,
  legend_args = NA,
  ...
)

\method{plot}{zvalues}(
  x,
  x_list = NULL,
  style = c("gray", "bw", "colored"),
  which = c("modulus", "phase", "nyquist", "real"),
  subfigure_main = NA,
  xlim = NA,
  ylim = NA,
  main = NA,
  ylab = NA,
  xlab = NA,
  legend = NULL,
  legend_args = NA,
  col = NA,
  type = "l",
  lty = "solid",
  lwd = 1,
  pch = 16,
  cex.points = 1,
  bg.points = "black",
  ...
)
}
\arguments{
\item{x}{\code{\link{pseries}} or \code{\link{zvalues}} object.}

\item{x_list}{(optional) list of additional \code{pseries} or \code{zvalues} objects.}

\item{xlim}{determine the axis limits of the subfigures.
E.g. \code{xlim = 'column'} means that all subfigures in a column
use the same x-axis limits. The parameter \code{xlim} may also
contain a 2-dimensional vector \code{c(x1,x2)}. In this case
all sub-figures use the given limits for the x-axis. Furthermore
the limits for the y-axis are computed based on the corresponding
data "subset".}

\item{ylim}{determine the axis limits of the subfigures.
E.g. \code{xlim = 'column'} means that all subfigures in a column
use the same x-axis limits. The parameter \code{xlim} may also
contain a 2-dimensional vector \code{c(x1,x2)}. In this case
all sub-figures use the given limits for the x-axis. Furthermore
the limits for the y-axis are computed based on the corresponding
data "subset".}

\item{main}{(character or \code{\link{expression}}) main title of the plot}

\item{xlab}{(character string or \code{\link{expression}}) label for the x-axis}

\item{ylab}{(character or \code{\link{expression}}) label for the y-axis}

\item{subfigure_main}{scalar or \code{(m x n)} matrix of type "character"
with the titles for the subfigures. If subfigure_main is a scalar
character string then the procedures creates a matrix of respective titles
by replacing the "place holders" '\code{i_}' and '\code{j_}'
with the respective row and column number.}

\item{parse_subfigure_main}{boolean. If \code{TRUE} then the titles for the subfigures
are parsed to \code{\link{expression}} before plotting. See also
\code{\link[grDevices]{plotmath}} on the usage of expressions for
plot annotations.}

\item{style}{(character string) determines the appearance of the plot
(background color of the plot regions, color and line style of the
grid lines, axis color, ...) See also \code{\link{style_parameters}}.}

\item{col}{vector of line colors}

\item{type}{vector of plot types. The following values are possible: 
"p" for points,
"l" for lines, 
"b" for both points and lines, 
"c" for empty points joined by lines,
"o" for overplotted points and lines, 
"s" and "S" for stair steps and
"h" for histogram-like vertical lines. 
\code{'n'} suppresses plotting.}

\item{lty}{vector of line types. Line types can either be specified as integers
(0=blank, 1=solid (default), 2=dashed, 3=dotted, 4=dotdash, 5=longdash, 6=twodash)
or as one of the character strings "blank", "solid", "dashed", "dotted", "dotdash",
"longdash", or "twodash", where "blank" uses ‘invisible lines’ (i.e., does not draw them).}

\item{lwd}{vector of line widths.}

\item{pch}{vector of plotting character or symbols. See \code{\link{points}} for possible
values.}

\item{cex.points}{vector of scales for the plotting symbols.}

\item{bg.points}{vector of fill color for the open plot symbols.}

\item{legend}{(character or \code{\link{expression}} vector). If \code{NULL} then
no legend is produced.}

\item{legend_args}{(optional) list with parameters for the legend. 
A legend title can be included with \code{legend_args = list(title = my_legend_title)}.
Note that the slots \code{x}, \code{y} are ignored and the legend is always put at the
right hand side of the plot. 
See also \code{\link[graphics]{legend}}.}

\item{...}{not used.}

\item{which}{(character string) what to plot (only used for \code{\link{zvalues}} 
objects). See details.}
}
\value{
The plot methods return (invisibly) a function, \code{subfig} say, which may be used to 
        add additional graphic elements to the subfigures. The call \code{opar = subfig(i,j)}
        creates a new (sub) plot at the (i,j)-the position with suitable margins and 
        axis limits. See the examples below.
}
\description{
Plot impulse and frequency response functions.
}
\details{
The parameter \code{which} determines what to plot in the case of \code{zvalues} objects.
\describe{
\item{modulus}{plot the moduli \code{abs(x[i,j])} versus frequencies \code{-Arg(z)/(2*pi)}.}
\item{phase}{plot the arguments \code{Arg(x[i,j])} versus frequencies \code{-Arg(z)/(2*pi)}.}
\item{nyquist}{plot the imaginary part \code{Im(x[i,j])} versus the real part \code{Re(x[i,j])}.}
\item{real}{plot the real part \code{Re(x[i,j])} versus the real part \code{Re(z)}. }
}
The choices \code{which = 'modulus'}, \code{which = 'phase'} and \code{which = 'nyquist'} 
are typically used for the case, where the rational matrix is evaluated on a (regular) grid 
of points on the unit circle ($|z|=1$). The choice \code{which = 'real'} is of course designed 
for the case that the z's are real numbers.

\code{NA} values for the (optional) parameters mean that the methods use some suitable default values. 
(For the frequency response plots, e.g. the labels for the x- and y-axis are chosen according to the 
parameter \code{which}.)

\code{NUll} values for the (optional) parameters mean that the respective graphic element is omitted.
E.g. \code{subfigure_main=NULL} skips the titles for the subfigures. 

The titles for the (m,n) subfigures are determined by the parameter \code{subfigure_main}. 
One may pass an \eqn{(m,n)} matrix of character strings to the procedure. Alternatively one may also 
provide an expression vector with a \code{dim} attribute such that {subfigure_main[i,j]} 
returns the expression which is to be used as title for the \eqn{(i,j)}-th subfigure. 
E.g. for a $(2,2)$ rational matrix one might use 
\preformatted{
subfigure_main = expression(Alpha, Beta, Gamma, Delta)
dim(subfigure_main) = c(2,2)
}

If \code{subfigure_main} is a scalar (character string or expression) then the procedures creates 
the respective titles by replacing the "place holders" \code{i_} and \code{j_} with the 
respective row and column number. See the examples below.

The parameter \code{xlim} determines the x-axis of the subfigures. 
If \code{xlim='subfigure'} then each subfigure gets its 
own x-axis (with different limits). For the case \code{xlim='column'} the subfigues in  
each column share a common x-axis (with common limits). The case \code{xlim = 'global'} 
means that all subfigures have the (x-) limits and the x-axis is only displayed 
for the subfigures in the last row. The parameter \code{ylim} is handled analogously. 

If more than one object is plotted (if the optional parameter \code{x_list} is not empty)
then a suitable legend may be added with the parameters \code{legend, legend_args}. 
Suppose that \eqn{k} objects are plotted, then \code{legend} should be a 
character (or expression) vector of length \code{k}. 

The "style" parameters \code{col, type, ..., bg.points} determine the appearance of the 
"lines" for each of the \code{k} objects. (If necessary the values are "recycled".)

These plot methods use the internal helper function \code{\link{plot_3D}}.
}
