% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_ratmat_plot_tools.R
\name{plot_3D}
\alias{plot_3D}
\title{Plot 3D Arrays}
\usage{
plot_3D(
  x,
  y,
  xlim = c("subfig", "column", "global"),
  ylim = c("subfig", "row", "global"),
  log = "",
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  subfigure_main = "(i_,j_)-th entry",
  parse_subfigure_main = FALSE,
  style = c("gray", "bw", "bw2", "colored"),
  col = NA,
  type = "l",
  lty = "solid",
  lwd = 1,
  pch = 16,
  cex.points = 1,
  bg.points = "black",
  legend = NULL,
  legend_args = NA
)
}
\arguments{
\item{x}{list of 3D-arrays (define the 'x'-values)}

\item{y}{list of 3D-arrays (define the 'y'-values)}

\item{xlim, ylim}{determine the axis limits of the subfigures.
E.g. \code{xlim = 'column'} means that all subfigures in a column
use the same x-axis limits. The parameter \code{xlim} may also
contain a 2-dimensional vector \code{c(x1,x2)}. In this case
all sub-figures use the given limits for the x-axis. Furthermore
the limits for the y-axis are computed based on the corresponding
data "subset".}

\item{log}{a character string which contains "x" if the x axis is to be logarithmic,
"y" if the y axis is to be logarithmic and "xy" or "yx" if both axes
are to be logarithmic.}

\item{main}{(character or \code{\link{expression}}) main title of the plot}

\item{xlab}{(character string or \code{\link{expression}}) label for the x-axis}

\item{ylab}{(character or \code{\link{expression}}) label for the y-axis}

\item{subfigure_main}{scalar or \code{(m x n)} matrix of type "character"
with the titles for the subfigures. If subfigure_main is a scalar
character string then the procedures creates a matrix of respective titles
by replacing the "place holders" '\code{i_}' and '\code{j_}'
with the respective row and column number.}

\item{parse_subfigure_main}{boolean. If \code{TRUE} then the titles for the subfigures
are parsed to \code{\link{expression}} before plotting. See also
\code{\link[grDevices]{plotmath}} on the usage of expressions for
plot annotations.}

\item{style}{(character string) determines the appearance of the plot
(background color of the plot regions, color and line style of the
grid lines, axis color, ...) See also \code{\link{style_parameters}}.}

\item{col}{vector of line colors}

\item{type}{vector of plot types. The following values are possible: 
"p" for points,
"l" for lines, 
"b" for both points and lines, 
"c" for empty points joined by lines,
"o" for overplotted points and lines, 
"s" and "S" for stair steps and
"h" for histogram-like vertical lines. 
\code{'n'} suppresses plotting.}

\item{lty}{vector of line types. Line types can either be specified as integers
(0=blank, 1=solid (default), 2=dashed, 3=dotted, 4=dotdash, 5=longdash, 6=twodash)
or as one of the character strings "blank", "solid", "dashed", "dotted", "dotdash",
"longdash", or "twodash", where "blank" uses ‘invisible lines’ (i.e., does not draw them).}

\item{lwd}{vector of line widths.}

\item{pch}{vector of plotting character or symbols. See \code{\link{points}} for possible
values.}

\item{cex.points}{vector of scales for the plotting symbols.}

\item{bg.points}{vector of fill color for the open plot symbols.}

\item{legend}{(character or \code{\link{expression}} vector). If \code{NULL} then
no legend is produced.}

\item{legend_args}{(optional) list with parameters for the legend. 
A legend title can be included with \code{legend_args = list(title = my_legend_title)}.
Note that the slots \code{x}, \code{y} are ignored and the legend is always put at the
right hand side of the plot. 
See also \code{\link[graphics]{legend}}.}
}
\value{
A \code{\link{function}} ("closure"), \code{subfig} say, which may be to add additional
        graphical elements to the subfigures. The call \code{opar = subfig(i,j)} sets up the
        coordinate system margins and figure coordinates such that one may add lines, text, points, ...
        to the \code{(i,j)}-th subfigure.
}
\description{
This internal helper function plots data arranged in 3-dimensional arrays.
The base application is as follows. Let x,y be two 3-dimensional arrays
with dimension c(m,n,l). Then \code{plot_3D} splits the device region into a
a matrix like array of (m times n) subfigures. In the (i,j)-th subfigure
\code{y[i,j,]} is plotted against \code{x[i,j,]}.
}
\details{
This function is mainly used by the plot methods, see \code{\link{plot methods}}.

For more information about the parameters \code{col, type, lty, ...,bg.points}
see \code{\link{plot.default}}.

This helper function only makes some very basic checks on the given parameters.
}
