% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_ratmat_classes.R
\name{polm}
\alias{polm}
\title{Constructor for Polynomial Matrices}
\usage{
polm(a)
}
\arguments{
\item{a}{either a (numeric or complex) vector, matrix or 3-D array. A vector is coerced to a scalar 
(i.e. \eqn{(1,1)}-dimensional) polynomial and a matrix gives a polynomial matrix of zero degree.}
}
\value{
An object of class \code{polm}.
}
\description{
This function was originally part of the R-package \strong{rationalmatrices}.
\cr
\code{polm} objects represent polynomial matrices
\deqn{a(z) = a_0 + a_1 z + \cdots + a_p z^p}{a(z) = a[0] + a[1]z + \dots + a[p]z^p}
If the matrix \eqn{a(z)} is an \eqn{(m,n)} dimensional polynomial matrix (i.e. the coefficients 
\eqn{a_i}{a[i]} are \eqn{(m,n)} dimensional real or complex valued matrices) then the 
\code{polm} object stores the coefficients in an \code{(m,n,p+1)} dimensional (numeric or complex) 
array together with a class attribute \code{c("polm","ratm")}. 
\cr
The constructor function \code{polm(a)} takes a (numeric or complex) vector, matrix or 
3-dimensional array and returns a \code{polm} object.
}
\details{
Any of the dimensions of the 3-dimensional array may also be zero. In particular, 
if the third dimension is zero, then the \code{polm} object is interpreted as the zero polynomial.

For important methods and functions for this class have a look at the "see also" section.
}
\examples{
# (1 x 1) polynomial matrix a(z) =  0 + 1z + 2z^2
polm(0:2)

# (2 x 3) polynomial matrix a(z) = a0 (degree is zero)
polm(diag(1, nrow = 2, ncol = 3))

# random (2 x 3) polynomial matrix a(z) = a0 + a1 z + a2 z^2 + a3 z^3 (degree = 3)
polm(array(stats::rnorm(2*3*4), dim = c(2,3,4)))

# random (2 x 1) polynomial matrix with complex coefficients (degree = 2)
a = polm(array(complex(real = stats::rnorm(2*1*3), 
               imaginary = stats::rnorm(2*1*3)), dim = c(2,1,3)))
is.polm(a)
dim(a)
str(a)
print(a, digits = 3)
}
\seealso{
\itemize{
   \item \code{\link{test_polm}} generates random polynomials. 
   \item checks: \code{\link{is.polm}}, \code{\link{is.miniphase}}, 
         and \code{\link{is.coprime}}. As a byproduct \code{is.coprime(a)} 
         computes the zeroes of a square polynomial matrix \eqn{a(z)}. 
   \item generic S3 methods: \code{\link{dim}}, 
         \code{\link{str}} and \code{\link{print}}.
   \item arithmetics: \code{\link{Ops.ratm}}, matrix multiplication \code{\link{\%r\%}}, 
         polynomial division \code{\%/\%}, 
         polynomial remainder \code{\%\%}, ...
   \item matrix operations: \code{\link{t.polm}}, 
         \code{\link{bind}}, \code{\link{[.polm}}, \code{\link{[<-.polm}}, ...
   \item \code{\link{degree}} returns the degree, \code{\link{col_end_matrix}} computes the 
         \emph{column end matrix} and \code{\link{prune}} "simplifies" a polynomial. 
   \item \code{\link{blaschkerize}} and \code{\link{transform_allpass}} may be used 
         to reflect zeroes of a polynomial matrix by multiplication
         with allpass rational matrices.
   \item normal forms: Hermite normal form \code{\link{hnf}}, Smith normal form \code{\link{snf}},  
                       column reduced form \code{\link{col_reduce}} and Wiener Hopf factorization 
                       \code{\link{whf}}.
   \item \code{\link{companion_matrix}}, \code{\link{zeroes}}, \code{\link{pseries}}, \code{\link{zvalues}}, ...  
}
}
