% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_ratmat_print_methods.R,
%   R/zz_rldm_print_methods.R
\name{print methods}
\alias{print methods}
\alias{print.polm}
\alias{print.lmfd}
\alias{print.armamod}
\alias{print.stspmod}
\alias{print.impresp}
\alias{print.autocov}
\alias{print.fevardec}
\alias{print.freqresp}
\alias{print.spectrald}
\title{Print Methods}
\usage{
\method{print}{polm}(
  x,
  digits = NULL,
  format = c("i|jz", "i|zj", "iz|j", "zi|j", "i|j|z", "character"),
  ...
)

\method{print}{lmfd}(
  x,
  digits = NULL,
  format = c("i|jz", "i|zj", "iz|j", "zi|j", "i|j|z", "character"),
  ...
)

\method{print}{armamod}(
  x,
  digits = NULL,
  format = c("i|jz", "i|zj", "iz|j", "zi|j", "i|j|z", "character"),
  ...
)

\method{print}{stspmod}(x, digits = NULL, ...)

\method{print}{impresp}(
  x,
  digits = NULL,
  format = c("i|jz", "i|zj", "iz|j", "zi|j", "i|j|z"),
  ...
)

\method{print}{autocov}(
  x,
  digits = NULL,
  format = c("i|jz", "i|zj", "iz|j", "zi|j", "i|j|z"),
  ...
)

\method{print}{fevardec}(
  x,
  digits = NULL,
  format = c("i|jz", "i|zj", "iz|j", "zi|j", "i|j|z"),
  ...
)

\method{print}{freqresp}(
  x,
  digits = NULL,
  format = c("i|jz", "i|zj", "iz|j", "zi|j", "i|j|z"),
  ...
)

\method{print}{spectrald}(
  x,
  digits = NULL,
  format = c("i|jz", "i|zj", "iz|j", "zi|j", "i|j|z"),
  ...
)
}
\arguments{
\item{x}{RLDM object, i.e. a \code{\link{armamod}}, \code{\link{stspmod}},
\code{\link{impresp}}, \code{\link{autocov}}, \code{\link{freqresp}},
\code{\link{spectrum}} or \code{\link{fevardec}} object.}

\item{digits}{(integer) if non \code{NULL} then correspondingly rounded numbers are printed,
see \code{\link{round}}.}

\item{format}{(character string) selects specific output formats. Note that
\code{\link{stsp}} and \code{\link{fevardec}} objects have no format option.
The option \code{'character'} is only implemented for (V)ARMA models.}

\item{...}{Further parameters are ignored.}
}
\value{
\code{invisible(x)}

\code{invisible(x)}
}
\description{
These functions were originally part of the R-package \strong{rationalmatrices}.

This function was originally part of the R-package \strong{RLDM}.
}
\examples{
# for polynomials six different print formats are implemented ###################
a = test_polm(dim = c(2,3), degree = 2, random = TRUE)

for (fmt in c("i|jz", "i|zj", "iz|j", "zi|j", "i|j|z", "character")) {
   cat('\nformat =', fmt, '\n')
   print(a, digits = 2, format = fmt)
}

# "empty" (2 x 0) polynomial matrix (degree = 2)
a = test_polm(dim = c(2,0), degree = 0)
print(a)

# random (2 x 1) polynomial matrix with complex coefficients (degree = 2)
a = polm(array(complex(real = stats::rnorm(2*1*3), 
                       imaginary = stats::rnorm(2*1*3)), dim = c(2,1,3)))
print(a, digits = 2)
\dontrun{
# the format option 'character' is only implemented for polynomials matrices 
# with real coefficients!
print(a, digits = 2, format = 'character')
}

# print a rational matrix in 'lmfd' form 
a = test_lmfd(dim = c(2,3), degrees = c(2,1))
print(a, digits = 2, format = 'character')

# for VARMA models six different print formats are implemented ###################
m = armamod(test_lmfd(dim = c(2,2), degrees = c(1,1)), sigma_L = diag(2))
print(m, digits = 2, format = "i|jz")
}
