% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_rldm_print_methods.R
\name{print_3D}
\alias{print_3D}
\title{Print 3D array}
\usage{
print_3D(
  a,
  digits = NULL,
  format = c("i|jz", "i|zj", "iz|j", "zi|j", "i|j|z", "character")
)
}
\arguments{
\item{a}{3-dimensional (numeric or complex) array}

\item{digits}{(integer) if non \code{NULL} then correspondingly rounded numbers
are printed, see \code{\link{round}}.}

\item{format}{format string}
}
\value{
invisible(NULL)
}
\description{
This function was originally part of the R-package \strong{RLDM}.
}
\keyword{internal}
