% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_ratmat_polm_methods.R
\name{prune}
\alias{prune}
\title{Prune Matrix Polynomial}
\usage{
prune(x, tol = sqrt(.Machine$double.eps), brutal = FALSE)
}
\arguments{
\item{x}{\code{\link{polm}} object.}

\item{tol}{Double. Tolerance paramater. Default set to \code{sqrt(.Machine$double.eps)}.}

\item{brutal}{Boolean. Default set to FALSE.
If TRUE, all small elements are set to zero (irrespective of whether they are leading
coefficients or not).}
}
\value{
A matrix polynomial, i.e. a \code{\link{polm}} object.
}
\description{
This function was originally part of the R-package \strong{rationalmatrices}.
\cr
Performs three steps to simplify a matrix polynomial.
\enumerate{
\item All leading coefficients where the absolute values of the real and the imaginary parts are
  less than or equal to \code{tol} are set to zero.
\item The zero leading coefficient matrices are dropped.
\item If all the absolute values of the imaginary parts of the coefficients are less than or
      equal to \code{tol} then the coefficients are set to real values.
}
Empty polynomial matrices (i.e. matrices with zero rows or columns) are set to polynomial of zero degree.
}
\examples{
x = polm(array(c(1,0,0,0,
                 0,1,0,0,
                 0,0,1,0,
                 0,0,0,1,
                 0,0,0,0), dim = c(2,2,5)) + 1e-4)
x
prune(x, tol = 1e-3)
prune(x, tol = 1e-3, brutal = TRUE)

# Case of complex variables:
x = x + complex(imaginary = 1e-5)
x
prune(x, tol = 1e-3)

# also works for constant matrix polynomials (i.e. matrices)
x = polm(array(0:3, dim = c(2,2,1))+1e-4)
x
prune(x, tol = 1e-3)

# empty polynomials are coerced to polynomials of degree zero 
x = polm(array(0, dim = c(0,2,5)))
x
prune(x)
}
