% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_ratmat_tools.R
\name{QL and LQ decomposition}
\alias{QL and LQ decomposition}
\alias{ql_decomposition}
\alias{lq_decomposition}
\title{QL and LQ Decomposition}
\usage{
ql_decomposition(x, ...)

lq_decomposition(x, ...)
}
\arguments{
\item{x}{Matrix.}

\item{...}{Other arguments for \code{\link{qr}}}
}
\value{
List with two elements:
\item{q}{Orthogonal matrix}
\item{l}{Lower triangular matrix with non-negative diagonal elements.}
}
\description{
This function was originally part of the R-package \strong{rationalmatrices}.
\cr
Returns the QL and LQ decomposition of a matrix with non-negative "diagonal" elements in the QL and LQ decompositions.
Only works if no column pivoting occurs in the QR decomposition \code{\link{qr}}.
}
\note{
Shouldn't export this function when publishing package!
}
\section{Implementation of the QL decomposition using the QR decomposition}{

MORE DOCU PLEASE
}

\examples{
# QL Decomposition #############

set.seed(1803)

# Tall matrix
x = matrix(stats::rnorm(5*3), 5, 3)
out = ql_decomposition(x)
all.equal(x, out$q \%*\% out$l)
all.equal(diag(3), t(out$q) \%*\% out$q)
out$l

# Wide matrix
x = matrix(stats::rnorm(5*3), 3, 5)
out = ql_decomposition(x)
all.equal(x, out$q \%*\% out$l)
all.equal(diag(3), t(out$q) \%*\% out$q)
out$l

# Square matrix
x = matrix(stats::rnorm(5*3), 3, 3)
out = ql_decomposition(x)
all.equal(x, out$q \%*\% out$l)
all.equal(diag(3), t(out$q) \%*\% out$q)
out$l

# LQ Decomposition #############

# Tall matrix
x = matrix(stats::rnorm(5*3), 5, 3)
out = lq_decomposition(x)
all.equal(x, out$l \%*\% out$q)
all.equal(diag(3), out$q \%*\% t(out$q))
out$l

# Wide matrix
x = matrix(stats::rnorm(5*3), 3, 5)
out = lq_decomposition(x)
all.equal(x, out$l \%*\% out$q)
all.equal(diag(3), out$q \%*\% t(out$q))
out$l

# Square matrix
x = matrix(stats::rnorm(5*3), 3, 3)
out = lq_decomposition(x)
all.equal(x, out$l \%*\% out$q)
all.equal(diag(3), out$q \%*\% t(out$q))
out$l

# reset seed
set.seed(NULL)
}
