% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_ratmat_plot_tools.R
\name{range_3D}
\alias{range_3D}
\title{Range of Values in a 3-D Array}
\usage{
range_3D(x, MARGIN = integer(0))
}
\arguments{
\item{x}{\code{(m,n,k)}-dimensional numeric array}

\item{MARGIN}{integer(0), 1, 2 or (1:2). determines the "margin" over which to compute the
range(s).}
}
\value{
\code{(m,n,2)}-dimensional array.
}
\description{
\code{range_3D} computes the range of values of a three-dimensional array.
}
\details{
If the input, \code{x} say, is an \code{(m,n,k)} dimensional array then \code{range_3D}
returns an \code{(m,n,2)} dimensional array, \code{m} say:

\itemize{
\item \code{MARGIN = c(1,2)}: \code{m[i,j,]} contains the range of values in \code{x[i,j,]}.
\item \code{MARGIN = 1}: \code{m[i,j,]} contains the range of values in \code{x[i,,]}.
\item \code{MARGIN = 2}: \code{m[i,j,]} contains the range of values in \code{x[,j,]}.
\item \code{MARGIN = integer(0)}: \code{m[i,j,]} contains the range of values in \code{x[,,]}.
}
}
\keyword{internal}
