% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_ratmat_extract_methods.R
\name{replace}
\alias{replace}
\alias{[<-.polm}
\title{Replace Parts of a Polynomial Matrix}
\usage{
\method{[}{polm}(x, i, j) <- value
}
\arguments{
\item{x}{\code{\link{polm}} object}

\item{i, j}{indices}

\item{value}{Either a \code{\link{polm}} object, or a vector/matrix/array which may be coerced to 
a \code{polm} object by \code{polm(value)}.}
}
\description{
This function was originally part of the R-package \strong{rationalmatrices}.
\cr
The assigment operation \code{x[,] <- value} for polynomial matrices works quite analogously 
to the assigment operation of "ordinary" matrices.
}
\details{
Note that "named" arguments are not supported (in order to simplify the coding).
}
\examples{
a = test_polm(dim = c(3,2), degree = 1)
a
a[FALSE] = 0   # no items to replace, a is not changed
a
a[lower.tri(matrix(0, nrow = 3, ncol = 2))] = 0 # set elements below the diagonal equal to zero
a
a[3,1] = c(1,-1) # set (3,1) element
a
a[1:2, 2:1] = c(0,1)
a
a[2, ] = test_polm(dim = c(1,2), degree = 4)
a
a[, 1] = test_polm(dim = c(2,1), degree = 4) # this gives a warning
a

\dontrun{
a[i=1] = 0   # named arguments are not supported
}  
}
