% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_llfactory.R
\name{replace_noise}
\alias{replace_noise}
\title{Replace noise parameters}
\usage{
replace_noise(params, params_new_noise)
}
\arguments{
\item{params}{vector of doubles. Deep system and noise parameters. Its noise parameters will be replaced by \code{params_new_noise}.}

\item{params_new_noise}{vector of doubles. Noise parameters as obtained from \link{get_ic}.}
}
\value{
Vector of deep parameters \code{params} whose noise-part is updated with rotation such that the shock components are independent.
}
\description{
Called in \link{create_results_list}.
\cr
Replaces unidentified estimate obtained from Gaussian estimation with independent components obtained from call to \link{get_ic}.
\cr
Note that this function makes only sense when shock distribution is Gaussian or Laplace
}
