% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_filltemplates_helpers_02.R
\name{sgt_skew_kurt}
\alias{sgt_skew_kurt}
\title{Skewness and kurtosis for SGT}
\usage{
sgt_skew_kurt(l, p, q)
}
\arguments{
\item{l}{double between -1 and 1. Parametrises skewness}

\item{p, q}{positive double.  parametrise kurtosis}
}
\value{
List with slots \code{skewness} ,\code{kurtosis}.
}
\description{
Calculates skewness and kurtosis for the SGT family of distributions with mean 0 and variance 1 and parameters \code{l}, \code{p}, \code{q}.
Existence of moments depends on parameter values of \code{p}, \code{q}.
See \url{https://www.jstor.org/stable/2634700} and \url{https://cran.r-project.org/web/packages/sgt/vignettes/sgt.pdf}.
}
\examples{
sgt_skew_kurt(-0.52,1.49,85616)
sgt_skew_kurt(0.15,1.92,7.89)
}
\seealso{
\link[sgt]{sgt}
}
