% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_ratmat_str_methods.R, R/zz_rldm_impresp.R,
%   R/zz_rldm_str_methods.R
\name{str methods}
\alias{str methods}
\alias{str.polm}
\alias{str.lmfd}
\alias{str.impresp}
\alias{str.armamod}
\title{Display the Structure of Objects}
\usage{
\method{str}{polm}(object, ...)

\method{str}{lmfd}(object, ...)

\method{str}{impresp}(object, ...)

\method{str}{armamod}(object, ...)
}
\arguments{
\item{object}{an object}

\item{...}{not used}
}
\value{
invisible(NULL)

invisible(NULL)
}
\description{
This function was originally part of the R-package \strong{rationalmatrices}.
\cr

This function was originally part of the R-package \strong{RLDM}.
}
