% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_ratmat_plot_tools.R
\name{style_parameters}
\alias{style_parameters}
\title{Style Parameters}
\usage{
style_parameters(style = c("bw", "bw2", "gray", "colored"))
}
\arguments{
\item{style}{character indicating the desired "style".}
}
\value{
A list with slots:
  \item{border.grid}{background color of the plot region}
  \item{col.grid,lty.grid,lwd.grid}{color, line style and line width of the grid lines}
  \item{col.axis}{axis color}
  \item{col.box,lwd.box}{color and linew idth  of the box surrounding the plot region}
  \item{bg.labels,border.labels}{background and border color for axis labels}
  \item{col.labels}{text color of the axis labels}
}
\description{
The plotting functions use a list of common style parameters.
}
\keyword{internal}
