% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_filltemplates_helpers_02.R
\name{unique_p_s}
\alias{unique_p_s}
\title{Global identifiability by choosing a canonical representative}
\usage{
unique_p_s(mat, sig = NULL)
}
\arguments{
\item{mat}{A matrix with ones on the diagonal (as obtained from the maximum likelihood procedure).}

\item{sig}{Vector containing the standard deviations of the i.i.d. shocks  (as obtained from the ML procedure).}
}
\value{
List with items \code{mat} and \code{sig}, containing the permuted and scaled versions of the inputs such that the identification scheme is satisfied.
  The diagonal elements of \code{mat} are again equal to one.
}
\description{
Obtaining the canonical representative as described in the identification scheme in Lanne, Meitz, Saikkonen (2017).
}
\examples{
mm <- matrix(c(1,3,2,1), 2)
ss <- c(1,1)

ll <- unique_p_s(mm, ss)
(m = ll[["mat"]])
(s = ll[["sig"]])
m \%*\% diag(s) \%*\% diag(s) \%*\% t(m)
mm \%*\% diag(ss) \%*\% diag(ss) \%*\% t(mm)

set.seed(1)
n = 3
mat = matrix(rnorm(n^2), n ,n)
unique_p_s(mat)
}
